+++ src/main/java/com/Travelrithm/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
         UserEntity findByName(String name);
         Optional<UserEntity> findByEmail(String email);
 
-        Integer userId(Integer userId);
+        Long userId(Integer userId);
 }
