+++ src/main/java/com/Travelrithm/security/jwt/JWTUtil.java
@@ -23,8 +23,8 @@ public class JWTUtil {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("email", String.class);
     }
 
-    public Integer getUserId(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("userId", Integer.class);
+    public Long getUserId(String token) {
+        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("userId", Long.class);
     }
     public String getNickname(String token){
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("nickname", String.class);
@@ -40,7 +40,7 @@ public class JWTUtil {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().getExpiration().before(new Date());
     }
 
-    public String createJwt(Integer userId, String email, String nickname, String role, Long expiredMs) {
+    public String createJwt(Long userId, String email, String nickname, String role, Long expiredMs) {
 
         return Jwts.builder()
                 .claim("userId",userId)
